Added more business partner and user features
